fix: allow auto-correct to update message text before sending message (#22278)
This change resolves an issue with the iOS auto-correct feature accidentally populating the chat text-input after sending the message. The solution involves us configuring the text-input and send button to wait for the iOS auto-correct feature to change the text before sending the message. This behaviour follows the same behaviour seen in similar chat apps on iOS like WhatsApp, Signal, and iMessages.