ci: allow Nix cache collect step to fail
Currently some paths in Nix store cause an error like this:
```
error: path '/nix/store/27fhcnb96303i45p78wmidnvdyklfr7s-qtbase-6c09620' is not valid
```
Which can have many causes. It's not fatal but it does stop the job.

We can still upload whatever did work and was saved to the store paths
file even if that error is thrown.

Signed-off-by: Jakub Sokołowski <jakub@status.im>