chore: remove reagent override because it is currently not effective (#22320)
This change removes an override for the reagent polyfill for requestAnimationFrame.

Initially, the codebase used an override to configure reagent to use a different polyfill implementation,
however the override is not currently effective because of how reagent references the polyfill.

In this case, the `fake-raf` def is used by the `next-tick` def inside reagent, but both of those
defs are evaluated before the override is set, meaning that the override does not take
effect since we're not re-evaluating the next-tick def again.

So for now, we have replaced the override with a comment explaining how we could correctly use the override in the future if needed.