chore: Remove traces of identity rings in the mobile app (#22360)
Removes the identity ring (colorful ring around avatars) from the mobile client.
status-go isn't going to be changed until status-desktop is updated as well.

Why: there's general consensus that the ring colors are a flawed mechanism
against impersonation. Further details in issue
https://github.com/status-im/status-mobile/issues/21743 and original discussion
in https://github.com/status-im/status-mobile/issues/20617. There's agreement
from the Design team that the feature should be removed and Figma is either
up-to-date without rings or is in the process of being updated.

We are not attempting to resolve the impersonation problem right now.
Eventually, we will return to the drawing board and come up with better
solutions.

Fixes https://github.com/status-im/status-mobile/issues/21743